Updated "Undo" and "AddLine" Functions
When adding a new line, the polynomial coefficients stay the same as the
previous line, and is overlayed on the graph. This allows for easy
comparison of changing coefficients, as Dr Thomson asked for (I only
just remembered that)
When undoing (aka removing the last line), the sliders and textboxes are
updated to show the current polynomial coefficients.